created a singlereview component to further modularize the app
