+++ client/components/Reviews.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { deleteReview, editReview } from '../store/thunks'
-import UpdateReview from './UpdateReview';
+import SingleReview from './SingleReview'
 
 class Reviews extends React.Component {
 
@@ -12,7 +11,7 @@ class Reviews extends React.Component {
     }
 
     render() {
-        const { reviews, onDelete, user } = this.props
+        const { reviews } = this.props
 
         return (
             !reviews.length ?
@@ -21,23 +20,11 @@ class Reviews extends React.Component {
                 <div>
                     <h4>Reviews</h4>
                     <ul>
-                        {reviews.map(review => {
-                            return (
-                                <div key={review.id}
-                                    style={{ border: '2px solid black', borderRadius: '5px' }}>
-                                    <span>Rating: {review.rating}</span>
-                                    <p>{review.text}</p>
-                                    {' '}
-                                    <button onClick={() => onDelete(review.productId, review.id)}>Delete Review</button>
-                                    {
-                                        user.id === review.userId || user.admin === true ?
-                                            <UpdateReview reviewId={review.id} productId={review.productId} />
-                                            :
-                                            null
-                                    }
-                                </div>
-                            )
-                        })}
+                        {
+                            reviews.map(review => (
+                                <SingleReview key={review.id} reviewId={review.id} />
+                            ))
+                        }
                     </ul>
                 </div>
         )
@@ -49,8 +36,4 @@ const mapStateToProps = ({ reviews, auth }) => ({
     user: auth.user
 })
 
-const mapDispatchToProps = (dispatch) => ({
-    onDelete: (productId, reviewId) => dispatch(deleteReview(productId, reviewId))
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(Reviews)
+export default connect(mapStateToProps)(Reviews)
